feat(*): 删除文件

+++ tool/src/main/java/cn/cunchang/DeleteFileToTrash.java
@@ -8,6 +8,7 @@ import java.io.IOException;
 
 /**
  * 删除指定文件夹
+ *
  * @author cunchang
  * @date 2020/03/15
  */
@@ -15,18 +16,17 @@ public class DeleteFileToTrash {
 
     // https://blog.csdn.net/Felix_Dreammaker/article/details/79180471
     public static void main(String[] args) {
-
+        
         FileUtils fileUtils = W32FileUtils.getInstance();
         boolean flag = fileUtils.hasTrash();
-
-        try {
-            fileUtils.moveToTrash(new File[]{new File("C:\\Users\\Administrator\\Desktop\\xxxxx.bmp")});
-        } catch (IOException e) {
-            e.printStackTrace();
+        if (flag) {
+            try {
+                fileUtils.moveToTrash(new File[]{new File("C:\\Users\\Administrator\\Desktop\\xxxxx.bmp")});
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
         }
 
     }
-
-
 
 }
